kusto: consolidating onto API version `2023-05-02`
